New auto report request system
Moderators or staff linked in the data/staffSync.json folder will be sent a special DM and will be able to setup a ban / kick / mute reason in watchlist without needing to lookup user IDs. This function automatically picks up the ban / kick / mute from the SCP server plugin and will accept a reason in DMs and add it to the watchlist. Usage is straightforward in DMs, just respond to the bot and it will interpret any message response as your reason. BEST FEATURE, you can edit any mistakes you make in a reason entry by just entering a whole new message or editing the last message you sent.