Improvements all around
I redid a bit of the code for the reason request system, so the DM reason requests should be more reliable and less likely to break. Keep in mind the discordAPI has a mind of its own..
I also added the date feature to watchlists so now each watchlist report added will now feature a saved date so you know when the watchlist report was added. Previous versions did not save this data, so old watchlist reports that are looked up will not have a date stamp printed in the discord embed.
I also noticed that Discord.js does a really fucking shitty job at error handling for websockets, and while future versions do a better job at this issue, I don't want to rewrite the code to make it compatible with the newer versions of Discord.js. That being said this change should help deal with a few more of the common websocket errors that crash the bot. However there are likely more errors I need to fish out and patch. I refuse to restart this bot every single day when you can just simple work the problems and make the program JUST work.